+++ neurocaps/_utils/_check_kwargs.py
@@ -10,6 +10,6 @@ def _check_kwargs(defaults, **kwargs):
     if kwargs:
         invalid_kwargs = {k: v for k, v in kwargs.items() if k not in plot_dict}
         if invalid_kwargs:
-            LG.info(f"Invalid kwargs arguments used and will be ignored {invalid_kwargs}.")
+            LG.info(f"The following invalid kwargs arguments used and will be ignored: {invalid_kwargs}")
 
     return plot_dict