+++ neurocaps/_utils/_timeseriesextractor_internals/_timeseriesextractorgetter.py
@@ -49,9 +49,9 @@ class _TimeseriesExtractorGetter:
 
     @subject_timeseries.setter
     def subject_timeseries(self, subject_dict):
-        error_message = ("Must be a valid pickle file/subject timeseries should contain a nested dictionary where the "
-                         "first level is the subject id, second level is the run number in the form of 'run-#', and the "
-                         "final level is the timeseries as a numpy array.")
+        error_message = ("A valid pickle file/subject timeseries should contain a nested dictionary where the "
+                         "first level is the subject id, second level is the run number in the form of 'run-#', and "
+                         "the final level is the timeseries as a numpy array.")
         if isinstance(subject_dict, str) and subject_dict.endswith(".pkl"):
             self._subject_timeseries = _convert_pickle_to_dict(subject_dict)
         elif isinstance(subject_dict, dict):